+++ modules/backend/traits/WidgetMaker.php
@@ -1,6 +1,7 @@
 <?php namespace Backend\Traits;
 
 use Lang;
+use Backend\Classes\FormField;
 use Backend\Classes\WidgetManager;
 use SystemException;
 
@@ -19,10 +20,10 @@ trait WidgetMaker
     /**
      * Makes a widget object with the supplied configuration file.
      * @param string $class Widget class name
-     * @param array $configuration An array of config.
+     * @param array $widgetConfig An array of config.
      * @return WidgetBase The widget object
      */
-    public function makeWidget($class, $configuration = [])
+    public function makeWidget($class, $widgetConfig = [])
     {
         $controller = property_exists($this, 'controller') && $this->controller
             ? $this->controller
@@ -34,6 +35,43 @@ trait WidgetMaker
             ]));
         }
 
-        return new $class($controller, $configuration);
+        return new $class($controller, $widgetConfig);
+    }
+
+    /**
+     * Makes a form widget object with the supplied form field and widget configuration.
+     * @param string $class Widget class name
+     * @param mixed $fieldConfig A field name, an array of config or a FormField object.
+     * @param array $widgetConfig An array of config.
+     * @return FormWidgetBase The widget object
+     */
+    public function makeFormWidget($class, $fieldConfig = [], $widgetConfig = [])
+    {
+        $controller = property_exists($this, 'controller') && $this->controller
+            ? $this->controller
+            : $this;
+
+        if (!class_exists($class)) {
+            throw new SystemException(Lang::get('backend::lang.widget.not_registered', [
+                'name' => $class
+            ]));
+        }
+
+        if (is_string($fieldConfig)) {
+            $fieldConfig = ['name' => $fieldConfig];
+        }
+
+        if (is_array($fieldConfig)) {
+            $formField = new FormField(
+                array_get($fieldConfig, 'name'),
+                array_get($fieldConfig, 'label')
+            );
+            $formField->displayAs('widget', $fieldConfig);
+        }
+        else {
+            $formField = $fieldConfig;
+        }
+
+        return new $class($controller, $formField, $widgetConfig);
     }
 }